Struct gstreamer_base::Adapter

pub struct Adapter(_, _);

This class is for elements that receive buffers in an undesired size. While for example raw video contains one image per buffer, the same is not true for a lot of other formats, especially those that come directly from a file. So if you have undefined buffer sizes and require a specific size, this object is for you.

An adapter is created with Adapter::new. It can be freed again with gobject::ObjectExt::unref.

The theory of operation is like this: All buffers received are put into the adapter using Adapter::push and the data is then read back in chunks of the desired size using Adapter::map/Adapter::unmap and/or Adapter::copy. After the data has been processed, it is freed using Adapter::unmap.

Other methods such as Adapter::take and Adapter::take_buffer combine Adapter::map and Adapter::unmap in one method and are potentially more convenient for some use cases.

For example, a sink pad's chain function that needs to pass data to a library in 512-byte chunks could be implemented like this:

static GstFlowReturn
sink_pad_chain (GstPad *pad, GstObject *parent, GstBuffer *buffer)
{
  MyElement *this;
  GstAdapter *adapter;
  GstFlowReturn ret = GST_FLOW_OK;

  this = MY_ELEMENT (parent);

  adapter = this->adapter;

  // put buffer into adapter
  gst_adapter_push (adapter, buffer);

  // while we can read out 512 bytes, process them
  while (gst_adapter_available (adapter) >= 512 && ret == GST_FLOW_OK) {
    const guint8 *data = gst_adapter_map (adapter, 512);
    // use flowreturn as an error value
    ret = my_library_foo (data);
    gst_adapter_unmap (adapter);
    gst_adapter_flush (adapter, 512);
  }
  return ret;
}

For another example, a simple element inside GStreamer that uses Adapter is the libvisual element.

An element using Adapter in its sink pad chain function should ensure that when the FLUSH_STOP event is received, that any queued data is cleared using Adapter::clear. Data should also be cleared or processed on EOS and when changing state from gst::State::Paused to gst::State::Ready.

Also check the GST_BUFFER_FLAG_DISCONT flag on the buffer. Some elements might need to clear the adapter after a discontinuity.

The adapter will keep track of the timestamps of the buffers that were pushed. The last seen timestamp before the current position can be queried with Adapter::prev_pts. This function can optionally return the number of bytes between the start of the buffer that carried the timestamp and the current adapter position. The distance is useful when dealing with, for example, raw audio samples because it allows you to calculate the timestamp of the current adapter position by using the last seen timestamp and the amount of bytes since. Additionally, the Adapter::prev_pts_at_offset can be used to determine the last seen timestamp at a particular offset in the adapter.

The adapter will also keep track of the offset of the buffers (GST_BUFFER_OFFSET) that were pushed. The last seen offset before the current position can be queried with Adapter::prev_offset. This function can optionally return the number of bytes between the start of the buffer that carried the offset and the current adapter position.

Additionally the adapter also keeps track of the PTS, DTS and buffer offset at the last discontinuity, which can be retrieved with Adapter::pts_at_discont, Adapter::dts_at_discont and Adapter::offset_at_discont. The number of bytes that were consumed since then can be queried with Adapter::distance_from_discont.

A last thing to note is that while Adapter is pretty optimized, merging buffers still might be an operation that requires a malloc and memcpy operation, and these operations are not the fastest. Because of this, some functions like Adapter::available_fast are provided to help speed up such cases should you want to. To avoid repeated memory allocations, Adapter::copy can be used to copy data into a (statically allocated) user provided buffer.

Adapter is not MT safe. All operations on an adapter must be serialized by the caller. This is not normally a problem, however, as the normal use case of Adapter is inside one pad's chain function, in which case access is serialized via the pad's STREAM_LOCK.

Note that Adapter::push takes ownership of the buffer passed. Use gst_buffer_ref before pushing it into the adapter if you still want to access the buffer later. The adapter will never modify the data in the buffer pushed in it.

Implements

glib::object::ObjectExt

Methods

impl Adapter

Important traits for Adapter
impl Read for Adapter

pub fn new() -> Adapter

Creates a new Adapter. Free with gobject::ObjectExt::unref.

Returns

a new Adapter

pub fn available(&self) -> usize

Gets the maximum amount of bytes available, that is it returns the maximum value that can be supplied to Adapter::map without that function returning None.

Returns

number of bytes available in self

pub fn available_fast(&self) -> usize

Gets the maximum number of bytes that are immediately available without requiring any expensive operations (like copying the data into a temporary buffer).

Returns

number of bytes that are available in self without expensive operations

pub fn clear(&self)

Removes all buffers from self.

pub fn copy_bytes(&self, offset: usize, size: usize) -> Option<Bytes>

Similar to gst_adapter_copy, but more suitable for language bindings. size bytes of data starting at offset will be copied out of the buffers contained in self and into a new glib::Bytes structure which is returned. Depending on the value of the size argument an empty glib::Bytes structure may be returned.

offset

the bytes offset in the adapter to start from

size

the number of bytes to copy

Returns

A new glib::Bytes structure containing the copied data.

pub fn distance_from_discont(&self) -> u64

pub fn dts_at_discont(&self) -> ClockTime

Get the DTS that was on the last buffer with the GST_BUFFER_FLAG_DISCONT flag, or GST_CLOCK_TIME_NONE.

Feature: v1_10

Returns

The DTS at the last discont or GST_CLOCK_TIME_NONE.

pub fn flush(&self, flush: usize)

Flushes the first flush bytes in the self. The caller must ensure that at least this many bytes are available.

See also: Adapter::map, Adapter::unmap

flush

the number of bytes to flush

pub fn get_buffer(&self, nbytes: usize) -> Option<Buffer>

Returns a gst::Buffer containing the first nbytes of the self, but does not flush them from the adapter. See Adapter::take_buffer for details.

Caller owns a reference to the returned buffer. gst_buffer_unref after usage.

Free-function: gst_buffer_unref

nbytes

the number of bytes to get

Returns

a gst::Buffer containing the first nbytes of the adapter, or None if nbytes bytes are not available. gst_buffer_unref when no longer needed.

pub fn get_buffer_fast(&self, nbytes: usize) -> Option<Buffer>

Returns a gst::Buffer containing the first nbytes of the self, but does not flush them from the adapter. See Adapter::take_buffer_fast for details.

Caller owns a reference to the returned buffer. gst_buffer_unref after usage.

Free-function: gst_buffer_unref

nbytes

the number of bytes to get

Returns

a gst::Buffer containing the first nbytes of the adapter, or None if nbytes bytes are not available. gst_buffer_unref when no longer needed.

pub fn get_buffer_list(&self, nbytes: usize) -> Option<BufferList>

Returns a gst::BufferList of buffers containing the first nbytes bytes of the self but does not flush them from the adapter. See Adapter::take_buffer_list for details.

Caller owns the returned list. Call gst_buffer_list_unref to free the list after usage.

nbytes

the number of bytes to get

Returns

a gst::BufferList of buffers containing the first nbytes of the adapter, or None if nbytes bytes are not available

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_list(&self, nbytes: usize) -> Vec<Buffer>

Returns a glib::List of buffers containing the first nbytes bytes of the self, but does not flush them from the adapter. See Adapter::take_list for details.

Caller owns returned list and contained buffers. gst_buffer_unref each buffer in the list before freeing the list after usage.

nbytes

the number of bytes to get

Returns

a glib::List of buffers containing the first nbytes of the adapter, or None if nbytes bytes are not available

pub fn masked_scan_uint32(
&self,
mask: u32,
pattern: u32,
offset: usize,
size: usize
) -> isize

Scan for pattern pattern with applied mask mask in the adapter data, starting from offset offset.

The bytes in pattern and mask are interpreted left-to-right, regardless of endianness. All four bytes of the pattern must be present in the adapter for it to match, even if the first or last bytes are masked out.

It is an error to call this function without making sure that there is enough data (offset+size bytes) in the adapter.

This function calls Adapter::masked_scan_uint32_peek passing None for value.

mask

mask to apply to data before matching against pattern

pattern

pattern to match (after mask is applied)

offset

offset into the adapter data from which to start scanning, returns the last scanned position.

size

number of bytes to scan from offset

Returns

offset of the first match, or -1 if no match was found.

Example:

// Assume the adapter contains 0x00 0x01 0x02 ... 0xfe 0xff

gst_adapter_masked_scan_uint32 (adapter, 0xffffffff, 0x00010203, 0, 256);
// -> returns 0
gst_adapter_masked_scan_uint32 (adapter, 0xffffffff, 0x00010203, 1, 255);
// -> returns -1
gst_adapter_masked_scan_uint32 (adapter, 0xffffffff, 0x01020304, 1, 255);
// -> returns 1
gst_adapter_masked_scan_uint32 (adapter, 0xffff, 0x0001, 0, 256);
// -> returns -1
gst_adapter_masked_scan_uint32 (adapter, 0xffff, 0x0203, 0, 256);
// -> returns 0
gst_adapter_masked_scan_uint32 (adapter, 0xffff0000, 0x02030000, 0, 256);
// -> returns 2
gst_adapter_masked_scan_uint32 (adapter, 0xffff0000, 0x02030000, 0, 4);
// -> returns -1

pub fn masked_scan_uint32_peek(
&self,
mask: u32,
pattern: u32,
offset: usize,
size: usize
) -> (isize, u32)

Scan for pattern pattern with applied mask mask in the adapter data, starting from offset offset. If a match is found, the value that matched is returned through value, otherwise value is left untouched.

The bytes in pattern and mask are interpreted left-to-right, regardless of endianness. All four bytes of the pattern must be present in the adapter for it to match, even if the first or last bytes are masked out.

It is an error to call this function without making sure that there is enough data (offset+size bytes) in the adapter.

mask

mask to apply to data before matching against pattern

pattern

pattern to match (after mask is applied)

offset

offset into the adapter data from which to start scanning, returns the last scanned position.

size

number of bytes to scan from offset

value

pointer to uint32 to return matching data

Returns

offset of the first match, or -1 if no match was found.

pub fn offset_at_discont(&self) -> u64

Get the offset that was on the last buffer with the GST_BUFFER_FLAG_DISCONT flag, or GST_BUFFER_OFFSET_NONE.

Feature: v1_10

Returns

The offset at the last discont or GST_BUFFER_OFFSET_NONE.

pub fn prev_dts(&self) -> (ClockTime, u64)

Get the dts that was before the current byte in the adapter. When distance is given, the amount of bytes between the dts and the current position is returned.

The dts is reset to GST_CLOCK_TIME_NONE and the distance is set to 0 when the adapter is first created or when it is cleared. This also means that before the first byte with a dts is removed from the adapter, the dts and distance returned are GST_CLOCK_TIME_NONE and 0 respectively.

distance

pointer to location for distance, or None

Returns

The previously seen dts.

pub fn prev_dts_at_offset(&self, offset: usize) -> (ClockTime, u64)

Get the dts that was before the byte at offset offset in the adapter. When distance is given, the amount of bytes between the dts and the current position is returned.

The dts is reset to GST_CLOCK_TIME_NONE and the distance is set to 0 when the adapter is first created or when it is cleared. This also means that before the first byte with a dts is removed from the adapter, the dts and distance returned are GST_CLOCK_TIME_NONE and 0 respectively.

offset

the offset in the adapter at which to get timestamp

distance

pointer to location for distance, or None

Returns

The previously seen dts at given offset.

pub fn prev_offset(&self) -> (u64, u64)

Get the offset that was before the current byte in the adapter. When distance is given, the amount of bytes between the offset and the current position is returned.

The offset is reset to GST_BUFFER_OFFSET_NONE and the distance is set to 0 when the adapter is first created or when it is cleared. This also means that before the first byte with an offset is removed from the adapter, the offset and distance returned are GST_BUFFER_OFFSET_NONE and 0 respectively.

Feature: v1_10

distance

pointer to a location for distance, or None

Returns

The previous seen offset.

pub fn prev_pts(&self) -> (ClockTime, u64)

Get the pts that was before the current byte in the adapter. When distance is given, the amount of bytes between the pts and the current position is returned.

The pts is reset to GST_CLOCK_TIME_NONE and the distance is set to 0 when the adapter is first created or when it is cleared. This also means that before the first byte with a pts is removed from the adapter, the pts and distance returned are GST_CLOCK_TIME_NONE and 0 respectively.

distance

pointer to location for distance, or None

Returns

The previously seen pts.

pub fn prev_pts_at_offset(&self, offset: usize) -> (ClockTime, u64)

Get the pts that was before the byte at offset offset in the adapter. When distance is given, the amount of bytes between the pts and the current position is returned.

The pts is reset to GST_CLOCK_TIME_NONE and the distance is set to 0 when the adapter is first created or when it is cleared. This also means that before the first byte with a pts is removed from the adapter, the pts and distance returned are GST_CLOCK_TIME_NONE and 0 respectively.

offset

the offset in the adapter at which to get timestamp

distance

pointer to location for distance, or None

Returns

The previously seen pts at given offset.

pub fn pts_at_discont(&self) -> ClockTime

Get the PTS that was on the last buffer with the GST_BUFFER_FLAG_DISCONT flag, or GST_CLOCK_TIME_NONE.

Feature: v1_10

Returns

The PTS at the last discont or GST_CLOCK_TIME_NONE.

pub fn take_buffer(&self, nbytes: usize) -> Option<Buffer>

Returns a gst::Buffer containing the first nbytes bytes of the self. The returned bytes will be flushed from the adapter. This function is potentially more performant than Adapter::take since it can reuse the memory in pushed buffers by subbuffering or merging. This function will always return a buffer with a single memory region.

Note that no assumptions should be made as to whether certain buffer flags such as the DISCONT flag are set on the returned buffer, or not. The caller needs to explicitly set or unset flags that should be set or unset.

Since 1.6 this will also copy over all GstMeta of the input buffers except for meta with the gst::MetaFlags::Pooled flag or with the "memory" tag.

Caller owns a reference to the returned buffer. gst_buffer_unref after usage.

Free-function: gst_buffer_unref

nbytes

the number of bytes to take

Returns

a gst::Buffer containing the first nbytes of the adapter, or None if nbytes bytes are not available. gst_buffer_unref when no longer needed.

pub fn take_buffer_fast(&self, nbytes: usize) -> Option<Buffer>

Returns a gst::Buffer containing the first nbytes of the self. The returned bytes will be flushed from the adapter. This function is potentially more performant than Adapter::take_buffer since it can reuse the memory in pushed buffers by subbuffering or merging. Unlike Adapter::take_buffer, the returned buffer may be composed of multiple non-contiguous gst::Memory objects, no copies are made.

Note that no assumptions should be made as to whether certain buffer flags such as the DISCONT flag are set on the returned buffer, or not. The caller needs to explicitly set or unset flags that should be set or unset.

This will also copy over all GstMeta of the input buffers except for meta with the gst::MetaFlags::Pooled flag or with the "memory" tag.

This function can return buffer up to the return value of Adapter::available without making copies if possible.

Caller owns a reference to the returned buffer. gst_buffer_unref after usage.

Free-function: gst_buffer_unref

nbytes

the number of bytes to take

Returns

a gst::Buffer containing the first nbytes of the adapter, or None if nbytes bytes are not available. gst_buffer_unref when no longer needed.

pub fn take_buffer_list(&self, nbytes: usize) -> Option<BufferList>

Returns a gst::BufferList of buffers containing the first nbytes bytes of the self. The returned bytes will be flushed from the adapter. When the caller can deal with individual buffers, this function is more performant because no memory should be copied.

Caller owns the returned list. Call gst_buffer_list_unref to free the list after usage.

nbytes

the number of bytes to take

Returns

a gst::BufferList of buffers containing the first nbytes of the adapter, or None if nbytes bytes are not available

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn take_list(&self, nbytes: usize) -> Vec<Buffer>

Returns a glib::List of buffers containing the first nbytes bytes of the self. The returned bytes will be flushed from the adapter. When the caller can deal with individual buffers, this function is more performant because no memory should be copied.

Caller owns returned list and contained buffers. gst_buffer_unref each buffer in the list before freeing the list after usage.

nbytes

the number of bytes to take

Returns

a glib::List of buffers containing the first nbytes of the adapter, or None if nbytes bytes are not available

impl Adapter

pub fn copy(&self, offset: usize, dest: &mut [u8])

Copies size bytes of data starting at offset out of the buffers contained in Adapter into an array dest provided by the caller.

The array dest should be large enough to contain size bytes. The user should check that the adapter has (offset + size) bytes available before calling this function.

dest

the memory to copy into

offset

the bytes offset in the adapter to start from

size

the number of bytes to copy

pub fn push(&self, buf: Buffer)

Adds the data from buf to the data stored inside self and takes ownership of the buffer.

buf

a gst::Buffer to add to queue in the adapter