Trait gstreamer_base::BaseSinkExt

pub trait BaseSinkExt {
    fn get_blocksize(&self) -> u32;

    fn get_drop_out_of_segment(&self) -> bool;

    fn get_last_sample(&self) -> Option<Sample>;

    fn get_latency(&self) -> ClockTime;

    fn get_max_bitrate(&self) -> u64;

    fn get_max_lateness(&self) -> i64;

    fn get_render_delay(&self) -> ClockTime;

    fn get_sync(&self) -> bool;

    fn get_throttle_time(&self) -> u64;

    fn get_ts_offset(&self) -> ClockTimeDiff;

    fn is_async_enabled(&self) -> bool;

    fn is_last_sample_enabled(&self) -> bool;

    fn is_qos_enabled(&self) -> bool;

    fn query_latency(&self) -> Option<(bool, bool, ClockTime, ClockTime)>;

    fn set_async_enabled(&self, enabled: bool);

    fn set_blocksize(&self, blocksize: u32);

    fn set_drop_out_of_segment(&self, drop_out_of_segment: bool);

    fn set_last_sample_enabled(&self, enabled: bool);

    fn set_max_bitrate(&self, max_bitrate: u64);

    fn set_max_lateness(&self, max_lateness: i64);

    fn set_qos_enabled(&self, enabled: bool);

    fn set_render_delay(&self, delay: ClockTime);

    fn set_sync(&self, sync: bool);

    fn set_throttle_time(&self, throttle: u64);

    fn set_ts_offset(&self, offset: ClockTimeDiff);

    fn wait(&self, time: ClockTime) -> (FlowReturn, ClockTimeDiff);

    fn wait_clock(&self, time: ClockTime) -> (ClockReturn, ClockTimeDiff);

    fn wait_preroll(&self) -> FlowReturn;

    fn get_property_async(&self) -> bool;

    fn set_property_async(&self, async: bool);

    fn get_property_enable_last_sample(&self) -> bool;

    fn set_property_enable_last_sample(&self, enable_last_sample: bool);

    fn get_property_qos(&self) -> bool;

    fn set_property_qos(&self, qos: bool);

    fn connect_property_async_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_blocksize_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_enable_last_sample_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_last_sample_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_max_bitrate_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_max_lateness_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_qos_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_render_delay_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_sync_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_throttle_time_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_ts_offset_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all BaseSink methods.

Implementors

BaseSink

Required Methods

fn get_blocksize(&self) -> u32

Get the number of bytes that the sink will pull when it is operating in pull mode.

Returns

the number of bytes self will pull in pull mode.

fn get_drop_out_of_segment(&self) -> bool

Checks if self is currently configured to drop buffers which are outside the current segment

Feature: v1_12

Returns

true if the sink is configured to drop buffers outside the current segment.

fn get_last_sample(&self) -> Option<Sample>

Get the last sample that arrived in the sink and was used for preroll or for rendering. This property can be used to generate thumbnails.

The gst::Caps on the sample can be used to determine the type of the buffer.

Free-function: gst_sample_unref

Returns

a gst::Sample. gst_sample_unref after usage. This function returns None when no buffer has arrived in the sink yet or when the sink is not in PAUSED or PLAYING.

fn get_latency(&self) -> ClockTime

Get the currently configured latency.

Returns

The configured latency.

fn get_max_bitrate(&self) -> u64

Get the maximum amount of bits per second that the sink will render.

Returns

the maximum number of bits per second self will render.

fn get_max_lateness(&self) -> i64

Gets the max lateness value. See BaseSinkExt::set_max_lateness for more details.

Returns

The maximum time in nanoseconds that a buffer can be late before it is dropped and not rendered. A value of -1 means an unlimited time.

fn get_render_delay(&self) -> ClockTime

Get the render delay of self. see BaseSinkExt::set_render_delay for more information about the render delay.

Returns

the render delay of self.

fn get_sync(&self) -> bool

Checks if self is currently configured to synchronize against the clock.

Returns

true if the sink is configured to synchronize against the clock.

fn get_throttle_time(&self) -> u64

Get the time that will be inserted between frames to control the maximum buffers per second.

Returns

the number of nanoseconds self will put between frames.

fn get_ts_offset(&self) -> ClockTimeDiff

Get the synchronisation offset of self.

Returns

The synchronisation offset.

fn is_async_enabled(&self) -> bool

Checks if self is currently configured to perform asynchronous state changes to PAUSED.

Returns

true if the sink is configured to perform asynchronous state changes.

fn is_last_sample_enabled(&self) -> bool

Checks if self is currently configured to store the last received sample in the last-sample property.

Returns

true if the sink is configured to store the last received sample.

fn is_qos_enabled(&self) -> bool

Checks if self is currently configured to send Quality-of-Service events upstream.

Returns

true if the sink is configured to perform Quality-of-Service.

fn query_latency(&self) -> Option<(bool, bool, ClockTime, ClockTime)>

Query the sink for the latency parameters. The latency will be queried from the upstream elements. live will be true if self is configured to synchronize against the clock. upstream_live will be true if an upstream element is live.

If both live and upstream_live are true, the sink will want to compensate for the latency introduced by the upstream elements by setting the min_latency to a strictly positive value.

This function is mostly used by subclasses.

live

if the sink is live

upstream_live

if an upstream element is live

min_latency

the min latency of the upstream elements

max_latency

the max latency of the upstream elements

Returns

true if the query succeeded.

fn set_async_enabled(&self, enabled: bool)

Configures self to perform all state changes asynchronously. When async is disabled, the sink will immediately go to PAUSED instead of waiting for a preroll buffer. This feature is useful if the sink does not synchronize against the clock or when it is dealing with sparse streams.

enabled

the new async value.

fn set_blocksize(&self, blocksize: u32)

Set the number of bytes that the sink will pull when it is operating in pull mode.

blocksize

the blocksize in bytes

fn set_drop_out_of_segment(&self, drop_out_of_segment: bool)

Configure self to drop buffers which are outside the current segment

Feature: v1_12

drop_out_of_segment

drop buffers outside the segment

fn set_last_sample_enabled(&self, enabled: bool)

Configures self to store the last received sample in the last-sample property.

enabled

the new enable-last-sample value.

fn set_max_bitrate(&self, max_bitrate: u64)

Set the maximum amount of bits per second that the sink will render.

max_bitrate

the max_bitrate in bits per second

fn set_max_lateness(&self, max_lateness: i64)

Sets the new max lateness value to max_lateness. This value is used to decide if a buffer should be dropped or not based on the buffer timestamp and the current clock time. A value of -1 means an unlimited time.

max_lateness

the new max lateness value.

fn set_qos_enabled(&self, enabled: bool)

Configures self to send Quality-of-Service events upstream.

enabled

the new qos value.

fn set_render_delay(&self, delay: ClockTime)

Set the render delay in self to delay. The render delay is the time between actual rendering of a buffer and its synchronisation time. Some devices might delay media rendering which can be compensated for with this function.

After calling this function, this sink will report additional latency and other sinks will adjust their latency to delay the rendering of their media.

This function is usually called by subclasses.

delay

the new delay

fn set_sync(&self, sync: bool)

Configures self to synchronize on the clock or not. When sync is false, incoming samples will be played as fast as possible. If sync is true, the timestamps of the incoming buffers will be used to schedule the exact render time of its contents.

sync

the new sync value.

fn set_throttle_time(&self, throttle: u64)

Set the time that will be inserted between rendered buffers. This can be used to control the maximum buffers per second that the sink will render.

throttle

the throttle time in nanoseconds

fn set_ts_offset(&self, offset: ClockTimeDiff)

Adjust the synchronisation of self with offset. A negative value will render buffers earlier than their timestamp. A positive value will delay rendering. This function can be used to fix playback of badly timestamped buffers.

offset

the new offset

fn wait(&self, time: ClockTime) -> (FlowReturn, ClockTimeDiff)

This function will wait for preroll to complete and will then block until time is reached. It is usually called by subclasses that use their own internal synchronisation but want to let some synchronization (like EOS) be handled by the base class.

This function should only be called with the PREROLL_LOCK held (like when receiving an EOS event in the ::event vmethod or when handling buffers in ::render).

The time argument should be the running_time of when the timeout should happen and will be adjusted with any latency and offset configured in the sink.

time

the running_time to be reached

jitter

the jitter to be filled with time diff, or None

Returns

gst::FlowReturn

fn wait_clock(&self, time: ClockTime) -> (ClockReturn, ClockTimeDiff)

This function will block until time is reached. It is usually called by subclasses that use their own internal synchronisation.

If time is not valid, no synchronisation is done and gst::ClockReturn::Badtime is returned. Likewise, if synchronisation is disabled in the element or there is no clock, no synchronisation is done and gst::ClockReturn::Badtime is returned.

This function should only be called with the PREROLL_LOCK held, like when receiving an EOS event in the BaseSinkClass.event() vmethod or when receiving a buffer in the BaseSinkClass.render() vmethod.

The time argument should be the running_time of when this method should return and is not adjusted with any latency or offset configured in the sink.

time

the running_time to be reached

jitter

the jitter to be filled with time diff, or None

Returns

gst::ClockReturn

fn wait_preroll(&self) -> FlowReturn

If the BaseSinkClass.render() method performs its own synchronisation against the clock it must unblock when going from PLAYING to the PAUSED state and call this method before continuing to render the remaining data.

If the BaseSinkClass.render() method can block on something else than the clock, it must also be ready to unblock immediately on the BaseSinkClass.unlock() method and cause the BaseSinkClass.render() method to immediately call this function. In this case, the subclass must be prepared to continue rendering where it left off if this function returns gst::FlowReturn::Ok.

This function will block until a state change to PLAYING happens (in which case this function returns gst::FlowReturn::Ok) or the processing must be stopped due to a state change to READY or a FLUSH event (in which case this function returns gst::FlowReturn::Flushing).

This function should only be called with the PREROLL_LOCK held, like in the render function.

Returns

gst::FlowReturn::Ok if the preroll completed and processing can continue. Any other return value should be returned from the render vmethod.

fn get_property_async(&self) -> bool

If set to true, the basesink will perform asynchronous state changes. When set to false, the sink will not signal the parent when it prerolls. Use this option when dealing with sparse streams or when synchronisation is not required.

fn set_property_async(&self, async: bool)

If set to true, the basesink will perform asynchronous state changes. When set to false, the sink will not signal the parent when it prerolls. Use this option when dealing with sparse streams or when synchronisation is not required.

fn get_property_enable_last_sample(&self) -> bool

fn set_property_enable_last_sample(&self, enable_last_sample: bool)

fn get_property_qos(&self) -> bool

fn set_property_qos(&self, qos: bool)

fn connect_property_async_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_blocksize_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_enable_last_sample_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_last_sample_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_max_bitrate_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_max_lateness_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_qos_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_render_delay_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_sync_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_throttle_time_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_ts_offset_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId